requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
drf-yasg==1.21.5
djangorestframework-simplejwt
